Move Kotlin file to src/main/kotlin directory
* Change javadoc generation from maven-javadoc-plugin to Dokka
* Update Xenon to 1.6.3
